% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fitbitshim}
\alias{fitbitshim}
\alias{fitbitshim-package}
\alias{read_fitbit}
\title{fitbitshim.}
\usage{
read_fitbit(key, secret, pilr_pt, start_date = "2013-01-01",
  end_date = "today")
}
\description{
A shim package for use with FitBit and PiLR API.
}

